requirement.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.5
defusedxml==0.7.1
deprecation==2.1.0
ecdsa==0.18.0
fastapi==0.104.0
future==0.18.3
h11==0.14.0
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
oic==1.6.1
packaging==23.2
pyasn1==0.5.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyjwkest==1.4.2
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==3.3.0
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
